fix/nix: fix stdout nulling for child processes
This causes EBADF (bad file descriptor) errors on Electron apps when
they output info to the console. These apps don't expect stdout to be
null, which is seemingly is the standard case for `Command` in stdlib.
This can be fixed by capturing output from stdout expicitly, using
`Command::output` instead of `Command::status`.